break initialState & constants into separate files, add public/src dir
