+++ public/index.html
@@ -12,8 +12,8 @@
         crossorigin="anonymous"></script>
     <script src="./lib/p5.gui.js"></script>
     <script src="./lib/quicksettings.js"></script>
-    <link rel="stylesheet" href="./style.css">
-    <script src="./app.js" type="module"></script>
+    <link rel="stylesheet" href="./src/style.css">
+    <script src="./src/app.js" type="module"></script>
 </head>
 
 <body>
