+++ public/src/constants.js
@@ -1,0 +1,1 @@
+export const dimensions = { width: 1920, height: 1000 };
